@@ -1,6 +1,6 @@
 module github.com/UltiRequiem/lorelai
 
-go 1.18
+go 1.22
 
 require (
 	github.com/UltiRequiem/chigo v1.3.0
