@@ -1,12 +1,13 @@
 package lorelai
 
 import (
+	"math/rand/v2"
 	"strings"
 )
 
 // Get a random Domain
 func Domain() string {
-	return trimSpaceAddDot(LoremWords(1)) + TLDS[Rng.Intn(len(TLDS))]
+	return trimSpaceAddDot(LoremWords(1)) + TLDS[rand.IntN(len(TLDS))]
 }
 
 // Get a random URL
